src/main/java/com/amadeus/nutrasoft/model/User.java
package com.amadeus.nutrasoft.model;

import java.util.Date;
import java.util.List;

public class User {
    private Integer id;
    private User coach;
    private String userType;
    private String name;
    private String gender;
    private Date birthdate;
    private String email;
    private String phone;
    private byte[] photo;
    private String summary;
    private String preferredLang;
    private byte status;
    private List<Plan> planList;

    private String photoName;
    private Short hrMax;
    private Byte donePct;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public User getCoach() {
        return coach;
    }

    public void setCoach(User coach) {
        this.coach = coach;
    }

    public String getUserType() {
        return userType;
    }

    public void setUserType(String userType) {
        this.userType = userType;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getGender() {
        return gender;
    }

    public void setGender(String gender) {
        this.gender = gender;
    }

    public Date getBirthdate() {
        return birthdate;
    }

    public void setBirthdate(Date birthdate) {
        this.birthdate = birthdate;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

    public byte[] getPhoto() {
        return photo;
    }

    public void setPhoto(byte[] photo) {
        this.photo = photo;
    }

    public String getSummary() {
        return summary;
    }

    public void setSummary(String summary) {
        this.summary = summary;
    }

    public String getPreferredLang() {
        return preferredLang;
    }

    public void setPreferredLang(String preferredLang) {
        this.preferredLang = preferredLang;
    }

    public byte getStatus() {
        return status;
    }

    public void setStatus(byte status) {
        this.status = status;
    }

    public List<Plan> getPlanList() {
        return planList;
    }

    public void setPlanList(List<Plan> planList) {
        this.planList = planList;
    }

    public String getPhotoName() {
        return photoName;
    }

    public void setPhotoName(String photoName) {
        this.photoName = photoName;
    }

    public Short getHrMax() {
        return hrMax;
    }

    public void setHrMax(Short hrMax) {
        this.hrMax = hrMax;
    }

    public Byte getDonePct() {
        return donePct;
    }

    public void setDonePct(Byte donePct) {
        this.donePct = donePct;
    }
}
